Test for Logging configuration updates. Remove unnecessary concurrent behaviour
